+++ layout/layout.php
@@ -23,7 +23,7 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later.
  */
 
- defined('MOODLE_INTERNAL') || die;
+defined('MOODLE_INTERNAL') || die;
 
 // Get the renderer for this plugin.
 $toolbox = \theme_foundation\toolbox::get_instance();