+++ Changes.md
@@ -10,6 +10,7 @@ Version 3.11.1.6 - TBR
 6. Fix 'Activity action menu cannot escape confines of #region-main and be displayed correctly'.
 7. Fix small issue with overridden 'build_action_menu_from_navigation' action menu code.
 8. Fix 'Message and notifications menu style'.
+9. Fix 'Maintenance layout not outputting doctype'.
 
 Version 3.11.1.5 - 13/02/22
 ---------------------------